pyproject.toml
[tool.poetry]
name = "report-aggregation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
streamlit = "^1.40.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
